refactor: move client code from test-harness to service integration tests
The test-harness crate should be purely server-side infrastructure.
It had a dependency on `databend-meta-client` solely for two client-creation
helpers (`make_grpc_client()` and `MetaSrvTestContext::grpc_client()`).
Move these to the service integration test module — the only consumer —
and drop the client dependency from test-harness.

Changes:
- Remove `databend-meta-client` dependency from `test-harness`
- Add free functions `make_grpc_client()` and `grpc_client()` in
  `crates/service/tests/it/tests/service.rs`
- Replace `tc.grpc_client()` method calls with `grpc_client(&tc)` across
  10 test files
