+++ crates/kvapi/src/lib.rs
@@ -15,3 +15,5 @@
 #![allow(clippy::uninlined_format_args)]
 
 pub mod kvapi;
+
+pub use kvapi::*;
