+++ crates/service/tests/it/grpc/metasrv_grpc_export.rs
@@ -25,6 +25,7 @@ use tokio::time::sleep;
 use tokio_stream::StreamExt;
 
 use crate::testing::meta_service_test_harness;
+use crate::tests::service::grpc_client;
 #[test(harness = meta_service_test_harness::<TokioRuntime, _, _>)]
 #[fastrace::trace]
 async fn test_export() -> anyhow::Result<()> {
@@ -37,7 +38,7 @@ async fn test_export() -> anyhow::Result<()> {
 
     let (tc, _addr) = crate::tests::start_metasrv::<TokioRuntime>().await?;
 
-    let client = tc.grpc_client().await?;
+    let client = grpc_client(&tc).await?;
 
     info!("--- upsert kv");
     {
