+++ crates/service/tests/it/grpc/t53_metasrv_grpc_snapshot_keys_layout.rs
@@ -24,6 +24,7 @@ use tokio::time::sleep;
 use tokio_stream::StreamExt;
 
 use crate::testing::meta_service_test_harness;
+use crate::tests::service::grpc_client;
 
 #[test(harness = meta_service_test_harness::<TokioRuntime, _, _>)]
 #[fastrace::trace]
@@ -33,7 +34,7 @@ async fn test_snapshot_keys_layout() -> anyhow::Result<()> {
     // Small delay to let service fully initialize
     sleep(Duration::from_millis(500)).await;
 
-    let client = tc.grpc_client().await?;
+    let client = grpc_client(&tc).await?;
 
     info!("--- Setup hierarchical test data");
     let test_keys = [
