+++ crates/test-harness/Cargo.toml
@@ -15,7 +15,6 @@ test = false
 anyhow = { workspace = true }
 databend-base = { workspace = true }
 databend-meta = { workspace = true }
-databend-meta-client = { workspace = true }
 databend-meta-runtime-api = { workspace = true }
 databend-meta-types = { workspace = true }
 fastrace = { workspace = true }
